refactor(onboarding): remove rollout contracts
