+++ apps/www/lib/routing/public/ownership.ts
@@ -41,11 +41,7 @@ function isPublicSurfacePath(
 }
 
 function isOnboardingPath(segments: readonly string[]) {
-  return (
-    segments.length === 0 ||
-    (segments.length === 1 &&
-      (segments[0] === "focus" || segments[0] === "role"))
-  );
+  return segments.length === 0;
 }
 
 function isUserPath(segments: readonly string[]) {
