+++ packages/backend/convex/_generated/api.d.ts
@@ -411,10 +411,6 @@ import type * as learningPreferences_impl from "../learningPreferences/impl.js";
 import type * as learningPreferences_mutations from "../learningPreferences/mutations.js";
 import type * as learningPreferences_program from "../learningPreferences/program.js";
 import type * as learningPreferences_queries from "../learningPreferences/queries.js";
-import type * as learningPrograms_mutations from "../learningPrograms/mutations.js";
-import type * as learningPrograms_queries from "../learningPrograms/queries.js";
-import type * as learningPrograms_selection from "../learningPrograms/selection.js";
-import type * as learningPrograms_spec from "../learningPrograms/spec.js";
 import type * as lib_attempts from "../lib/attempts.js";
 import type * as lib_effect from "../lib/effect.js";
 import type * as lib_helpers_auth from "../lib/helpers/auth.js";
@@ -432,10 +428,8 @@ import type * as lib_validators_vv from "../lib/validators/vv.js";
 import type * as notifications_mutations from "../notifications/mutations.js";
 import type * as notifications_queries from "../notifications/queries.js";
 import type * as onboarding_impl from "../onboarding/impl.js";
-import type * as onboarding_migrate from "../onboarding/migrate.js";
 import type * as onboarding_mutations from "../onboarding/mutations.js";
 import type * as onboarding_queries from "../onboarding/queries.js";
-import type * as onboarding_removeLegacySelection from "../onboarding/removeLegacySelection.js";
 import type * as onboarding_spec from "../onboarding/spec.js";
 import type * as onboarding_values from "../onboarding/values.js";
 import type * as polyfills from "../polyfills.js";
@@ -994,10 +988,6 @@ declare const fullApi: ApiFromModules<{
   "learningPreferences/mutations": typeof learningPreferences_mutations;
   "learningPreferences/program": typeof learningPreferences_program;
   "learningPreferences/queries": typeof learningPreferences_queries;
-  "learningPrograms/mutations": typeof learningPrograms_mutations;
-  "learningPrograms/queries": typeof learningPrograms_queries;
-  "learningPrograms/selection": typeof learningPrograms_selection;
-  "learningPrograms/spec": typeof learningPrograms_spec;
   "lib/attempts": typeof lib_attempts;
   "lib/effect": typeof lib_effect;
   "lib/helpers/auth": typeof lib_helpers_auth;
@@ -1015,10 +1005,8 @@ declare const fullApi: ApiFromModules<{
   "notifications/mutations": typeof notifications_mutations;
   "notifications/queries": typeof notifications_queries;
   "onboarding/impl": typeof onboarding_impl;
-  "onboarding/migrate": typeof onboarding_migrate;
   "onboarding/mutations": typeof onboarding_mutations;
   "onboarding/queries": typeof onboarding_queries;
-  "onboarding/removeLegacySelection": typeof onboarding_removeLegacySelection;
   "onboarding/spec": typeof onboarding_spec;
   "onboarding/values": typeof onboarding_values;
   polyfills: typeof polyfills;
